refactoring interactive dashboard code into more manageable modules
